requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
streamlit==1.49.1
requests==2.32.3
reportlab==4.2.2
langchain==0.3.7
openai==1.51.2
python-dotenv==1.0.1
huggingface-hub==0.26.2
fpdf==1.7.2
markdown2==2.5.4
newspaper3k==0.2.8
google-search-results==2.4.2   # for SerpAPI
google-generativeai==0.7.2     # for Gemini
